backend/requirements.txt
customtkinter
requests
beautifulsoup4
pillow
validators
python-docx
pywinstyles
fastapi
uvicorn
python-multipart
